从 webui.py 中拆分出 ui/recording.py
